+++ react/src/Grid.tsx
@@ -1,0 +1,18 @@
+import { useState, useEffect } from 'react';
+import './assets/bootstrap/bootstrap.min.css';
+import './assets/bootstrap/bootstrap.bundle.min'
+import './Grid.css';
+import './index.css';
+
+function Grid() {
+
+    return (
+        <>
+            <div className="content">
+                
+            </div>
+        </>
+    )
+}
+
+export default Grid;