+++ react/src/Home.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { toast, ToastContainer, Flip } from 'react-toastify';
 import './assets/bootstrap/bootstrap.min.css';
@@ -6,11 +6,14 @@ import './assets/bootstrap/bootstrap.bundle.min'
 import './Home.css';
 import './index.css';
 
+import GridSystem from './GridView.tsx';
+
 function Home() {
     const [user, setUser] = useState<user | null>(null);
     const [projects, setProjects] = useState<project[]>([]);
     const [selectedProject, setSelectedProject] = useState<project | null>(null);
     const navigate = useNavigate();
+    const gridRef = useRef<any>(null);
 
     function HandleProjectChange(id : number) : void {
         const proj : project | undefined = projects.find((p : project) => p.id == id);
@@ -19,7 +22,14 @@ function Home() {
         }
     }
 
+    function AddWidget() : void {
+        if (gridRef.current) {
+            gridRef.current.AddWidget();
+        }
+    }
+
     useEffect(() => {
+        // Fetches the currently logged in user
         async function GetUser() {
             const res : response = await window.api.GetCurrentUser();
             if (res.success && res.data) {
@@ -36,6 +46,7 @@ function Home() {
     }, []);
 
     useEffect(() => {
+        // Fetches the user's projects
         async function GetProjects(username : string) {
             const res : response = await window.api.GetProjects(username);
             if (res.success) {
@@ -55,26 +66,34 @@ function Home() {
         <ToastContainer limit={1} transition={Flip} />
         <div className="home-container">
             <div className="sidebar">
-                
+                <button className="user-info d-flex align-items-center">
+                    <img className="mt-2 ms-2" src="/default_pfp.png"></img>
+                    {user ? <h5 className="ms-2">{user.f_name} {user.l_name}</h5> : <h5>Loading...</h5>}
+                </button>
             </div>
             <div className="navbar">
                 <div className="dropdown">
                     <button type="button" className="btn btn-primary dropdown-toggle" data-bs-toggle="dropdown">
-                        {selectedProject ? selectedProject.name : "Select project"}
+                        {selectedProject ? selectedProject.name : "Select Palette"}
                     </button>
                     <ul className="dropdown-menu">
-                        {projects.map((proj : project, index : number) => (
-                            <li key={index}>
-                                <a id={proj.id.toString()} className="dropdown-item" href="#" 
-                                    onClick={() => HandleProjectChange(proj.id)}>{proj.name}
+                        {projects.length ? (projects.map((proj : project, index : number) => (
+                            <li key={proj.id}>
+                                <a className="dropdown-item" href="#" onClick={() => HandleProjectChange(proj.id)}>
+                                    {proj.name}
                                 </a>
                             </li>
-                        ))}
+                        ))) : (<li>
+                                <h6 className="dropdown-item">
+                                    No projects found. Create one below
+                                </h6>
+                              </li>)}
                     </ul>
+                    <button onClick={()=>AddWidget()}>New widget</button>
                 </div>
             </div>
             <div className="content">
-
+                <GridSystem ref={gridRef}/>
             </div>
         </div>
         </>